Remove circular logic but use of delay_mem is changed
When the up_val stops being sent the down valid will still be active on. Thus
if the 'dn_val' of one delay_mem is being used as a 'up_val' for another
delay_mem it would keep pushing the static value into the 2nd delay_mem module.
To prevent this the first 'up_val' signal must be used to AND all delay_mem
'up_val's.

Thus the 'delay' module works the same after this change but the delay_mem does
not.